Move the oauth details
