+++ portal/__init__.py
@@ -1,7 +1,6 @@
-from flask import Flask, url_for, redirect
-from authlib.flask.client import OAuth
+from flask import Flask
 
-from . import main
+from . import main, auth
 
 
 def create_app(test_config=None):
@@ -19,30 +18,6 @@ def create_app(test_config=None):
         app.config.from_mapping(test_config)
 
     app.register_blueprint(main.blueprint)
-
-    oauth = OAuth(app)
-    oauth.register(
-        'globus',
-        authorize_url='https://auth.globus.org/v2/oauth2/authorize',
-        client_id='12518f0d-4594-4632-8c4c-a6839024d238',
-        client_secret='gEfGGE09nMMjwZxWafL+2/M3UqcGl9czSL72H+O1xuU=',
-        state='placeholder'
-    )
-
-    @app.route('/auth/login')
-    def login():
-        redirect_uri = url_for('complete', _external=True)
-        return oauth.globus.authorize_redirect(redirect_uri)
-
-    # Redirect URL is registered with Globus:
-    # Hostname and path are required to exactly match their records,
-    # so we can't change this without re-registering.
-    @app.route('/auth/complete/globus/')
-    def complete():
-        # token = oauth.globus.authorize_access_token()
-        # resp = oauth.globus.get('user')
-        # profile = resp.json()
-        # do something with the token and profile
-        return redirect('/')
+    auth.add_auth(app)
 
     return app
